refactor(remediationorchestrator): extract terminal_transitions.go from reconciler.go
Phase 3 file split: moves errPhaseConflict, transitionToInheritedCompleted,
transitionToCompletedWithoutVerification, transitionToInheritedFailed,
handleBlocked, transitionPhase, VerificationDeadlineBuffer,
transitionToVerifying, transitionToFailed, handleGlobalTimeout, and
createEffectivenessAssessmentIfNeeded into their own file. errPhaseAlreadySet
stays in reconciler.go (used only by Reconcile). Pure structural move:
376 Ginkgo specs still pass. This completes the reconciler.go split
(3435 -> 1023 lines across 7 files).
